Added EPICS support for Cygwin and fixed a few EPICS-related glitches found by the Cygwin C compiler.
git-svn-id: http://svn.csrri.iit.edu/mx/trunk@1576 42547f41-83fc-0310-9aef-9668e019ab9e

+++ libMx/d_pmactc.c
@@ -519,7 +519,7 @@ mxd_pmac_tc_motor_motor_is_busy( MX_MOTOR *motor )
 	MX_DEBUG( 2,
 		("%s: motor '%s' final motion_state = %ld, motor->busy = %d",
 		fname, motor->record->name, pmac_tc_motor->motion_state,
-		motor->busy ));
+		(int) motor->busy ));
 
 	return MX_SUCCESSFUL_RESULT;
 }
